Removing deprecated models
Models 1-8 defined in the models.py file have been deprecated for a
while.  Given recent changes in pypm, most of these models don't work.
These models are being deleted here to cleanup pypm, though they are
presereved in earlier releases for reference.

Similarly, the models_new.py file is being deleted, which was our
first attempt to rework the setup of models.  But this work was
never completed.