Adding revised compact constraint
This refines the logic to provide a tighter representation when
all predecessor activities are scheduled before the time horizon.

Baselines needed to be updated to reflect the value of a new
decision variable.  z_pre[j] is zero if all of the predecessor
activities are scheduled before the time horizon.  And one otherwise.