Renaming models
This is a major change to model names.  The goal here is to

1. Preserve backwards compatibility
2. Allow for new default values of Gamma
3. Introduce more expressive names

This change was motivated by the observation that Gamma=None was the
default for GSF and GSF-ED, which amounts to allowing infinitely
many gaps in the execution of each activity.  This weakens the
formulation, and it is inconsistent with recent practices with pypm
applications to filter out un-interesting timesteps (and days).
Further, filtering out un-interesting timesteps also improves
the solve time.

This commit changes the default behavior of supervised and
unsupervised optimization to use the model
UnrestrictedMatches_VariableLengthActivities.  This model
differs from the previous default, GSF, in that the Gamma value is
set to zero by default.  I have seen some evidence that this change
does, in fact, provide better predictions for unsupervised
optimization (using tabu search) when using the default model.  Even so,
the use of XSF (a.k.a. UnrestrictedMatches_FixedLengthActivities)
is recommended because that model is easier to solve.